Product Pages Check
After learning how to utilize the useEffect() hook in a neat way, aka. firing only when the location (url) changes, I've managed to utilize local state which dynamically renders the styling of the marketplace container. For the general marketplace it scrolls, for the product page it is fixed and their will only be one small scroll section which is the configurable item selection page...

Now to finish styling the Links from react-router-dom because it utilizes the link html tag (which comes with default styling...) Minor but major success and fun learning curve.